+++ backend/canisters/community/impl/src/updates/update_bot.rs
@@ -22,11 +22,13 @@ fn update_bot_impl(args: Args, state: &mut RuntimeState) -> OCResult {
         return Err(OCErrorCode::InitiatorNotAuthorized.into());
     }
 
+    let now = state.env.now();
+
     if !state.data.update_bot(
         args.bot_id,
         args.granted_permissions.clone(),
         args.granted_autonomous_permissions.clone(),
-        state.env.now(),
+        now,
     ) {
         return Err(OCErrorCode::BotNotFound.into());
     }
@@ -45,6 +47,7 @@ fn update_bot_impl(args: Args, state: &mut RuntimeState) -> OCResult {
             granted_autonomous_permissions: args.granted_autonomous_permissions.unwrap_or_default(),
         })),
         recipients: vec![args.bot_id],
+        timestamp: now,
     }));
 
     handle_activity_notification(state);