+++ backend/canisters/user/impl/src/updates/c2c_install_bot.rs
@@ -68,6 +68,7 @@ fn c2c_install_bot_impl(args: Args, state: &mut RuntimeState) -> OCResult {
             granted_autonomous_permissions: args.granted_autonomous_permissions.unwrap_or_default(),
         })),
         recipients: vec![args.bot_id],
+        timestamp: now,
     }));
 
     Ok(())