+++ backend/canisters/user/impl/src/updates/update_bot.rs
@@ -58,6 +58,7 @@ fn update_bot_impl(args: Args, state: &mut RuntimeState) -> OCResult {
             granted_autonomous_permissions: args.granted_autonomous_permissions.unwrap_or_default(),
         })),
         recipients: vec![args.bot_id],
+        timestamp: now,
     }));
 
     Ok(())